Refactoring : let the fun begin...
Basic concept
1) A profile instantiates a person
2) A person, in turn, is gains dimension through value added analysis
provided by plug-in modules; i.e., to start, a 'kua' module which
provides
a) kua number
b) group
c) auspicious direction
all based on the person's profile data...

Big take away is that profile data is one thing
And added value analysis based on that data is another ... originally,
the idea was that feng shui-based kua analysis was a kind of profile,
which it turns out is a mistake: it adds a dimension to the person that
is secondary and which is based on primary data, specifically gender and
birth date.  Loosely considered, it could be spoken of as a kind of
'profile', but in architectural terms, it doesn't work well as such.

Problems: code is broken for the time being - namespace issue w/ mixin