Refactoring: code clean-up
1) Kua : details method broken down into several discrete jobs (method
itself removed):
a. set_birth_year
b. set_last_two_digits_birth_year
c. set_kua_number
d. set_group
e. set_auspicious_direction

The above are run under a function/method kua_init which is invoked when
Person is instantiated.

At the moment, code is broken due to a logic error that is somewhat
mysterious -- on the face of it, it's quite straightforward what should
happen here: anything on the closed interval [121,221] gets edge-case
treatment.  The treatment of everything else is academic.  In any case,
a no-brainer logic test is failing inexplicably, and can't figure out
why ... the short-term outfall is that a nil exception is being thrown
on an uninstantiated  variable pointing to an array ... Oh well.