Reimplementation: main algorithm re-engineer
Up till now used an ugly kludge to compare birthdates against lines in
the sand, i.e., new moon et al.  In the process of replacing with a
cleaner implementation by leveraging Ruby's built-in Date class, which I
originally eschewed b/c of some limitations that I don't even remember
what were.

Tests are running at 99.9%, w/ the only error an edge case test running
against the to-be-phased-out old implementation.

A first pass of the new implementation was added to the utilities module
under dates_to_compare -- a placeholder name that will be rechristened
w/ something better/more appropriate.  For the time being however, w/
the exception of a missing elsif clause that compares the birth date in
question with a calculated new year's date, the re-engineer is fairly
complete.  Still not sure what to do about edge cases where, for
example, birth date and new year's fall on the same day, or, where the
calculation of new year's falls short and above condition holds ... at
the moment the calculation of new year's is a fuzzy algorithm that would
probably benefit from programmatic deductive analysis, i.e., how
consistently is my algo nailing it and, otherwise, what's the margin of
error: so far, I've only found one-off errors, which seems to suggest
that granted not perfect, the algo is mostly on target...